Working on method for converting unix seconds to formatted date - need to get day of week. Method not being accessed. Added weather icons to drawable folder for displaying on fragment based on icon API value.

+++ app/src/main/java/com/epicodus/weatherpit/services/ForecastService.java
@@ -53,7 +53,7 @@ public class ForecastService {
                 JSONArray dailySummaryJSON = forecastServiceJSON.getJSONObject("daily").getJSONArray("data");
                 for (int i = 0; i < dailySummaryJSON.length(); i++) {
                     JSONObject summaryForecastJSON = dailySummaryJSON.getJSONObject(i);
-                    Integer time = summaryForecastJSON.getInt("time");
+                    Long time = summaryForecastJSON.getLong("time");
                     String summary = summaryForecastJSON.getString("summary");
                     String icon = summaryForecastJSON.getString("icon");
                     Double minTemp = summaryForecastJSON.getDouble("temperatureMin");
